Fix server.js
